+++ test/firestore_test.dart
@@ -1,5 +1,6 @@
 @TestOn('browser')
 import 'package:firebase/firebase.dart' as fb;
+import 'package:firebase/firebase_firestore.dart' as store;
 import 'package:firebase/src/assets/assets.dart';
 import 'package:test/test.dart';
 
@@ -24,5 +25,21 @@ void main() {
       await app.delete();
       app = null;
     }
+  });
+
+  group("Firestore", () {
+    store.Firestore firestore;
+
+    setUp(() {
+      firestore = fb.firestore();
+    });
+
+    group("instance", () {
+      test("App exists", () {
+        expect(firestore, isNotNull);
+        expect(firestore.app, isNotNull);
+        expect(firestore.app.name, fb.app().name);
+      });
+    });
   });
 }